requirements.txt
streamlit==1.37.1
pyyaml==6.0.2
pandas==2.2.2